levermann: fallback to data from previous year if it doesn't exist
In the first months of a new year, some operating figures for the last and
current year might not be known yet and only be released with the
following quarterly report.

Therefore if the equity_ratio of ebit_margin for the year is not
available, try to use the values of the previous year instead.