levermann: store analysis result in LevermannResult object
Store the result of an evaluation in a LevermannResult object together
with a timestamp.
Those are stored in a result in the Levermann object.

This enables to keep historic criteria dating and scores.
Previously the per criteria score and values were stored in the object.
Only historic Levermann scores and the date was in a list.