levermann: don't store result if old result is still up to date
Only store the evaluation result if the score changed or the previous
result is older than a week.

As part of this the creation of the __str__() output for the result was
moved to the LevermannResult() objects, where it belongs.
This was required to have proper debug output when the result isn't
stored.

Previously Levermann.__str__() was constructing the string for the
latest levermann analysis result.